+++ Test/EnemyBoardTest.java
@@ -1,3 +1,7 @@
+package Test;
+import Main.*;
+
+import org.junit.Assert;
 import org.junit.Test;
 import static org.junit.Assert.*;
 
@@ -11,8 +15,8 @@ public class EnemyBoardTest {
     @Test
     public void test_Constructor(){
         EnemyBoard e = new EnemyBoard();
-        assertEquals("Expected 17 ship elements",17,e.getNumberOfShipElements());
-        assertEquals("Expected 5 ships placed.",5,e.getShips().length);
+        Assert.assertEquals("Expected 17 ship elements",17,e.getNumberOfShipElements());
+        Assert.assertEquals("Expected 5 ships placed.",5,e.getShips().length);
     }
 
     @Test
@@ -58,10 +62,10 @@ public class EnemyBoardTest {
         }
 
 
-        assertEquals("Expected a ship at "+rowHasShip+","+colHasShip,3,e.getBoardElement(rowHasShip,colHasShip));
+        Assert.assertEquals("Expected a ship at "+rowHasShip+","+colHasShip,3,e.getBoardElement(rowHasShip,colHasShip));
         e.fire(rowHasShip,colHasShip);
-        assertEquals("Expected a hit ship at "+rowHasShip+","+colHasShip,2,e.getBoardElement(rowHasShip,colHasShip));
-        assertEquals("Should have 16 ship elements.",16,e.getNumberOfShipElements());
+        Assert.assertEquals("Expected a hit ship at "+rowHasShip+","+colHasShip,2,e.getBoardElement(rowHasShip,colHasShip));
+        Assert.assertEquals("Should have 16 ship elements.",16,e.getNumberOfShipElements());
 
         /*
                  PLAYER BOARD
@@ -104,11 +108,11 @@ public class EnemyBoardTest {
             }
         }
 
-        assertEquals("Should have 17 ship elements.",17,e.getNumberOfShipElements());
-        assertEquals("Expected no ship at "+rowNoShip+","+colNoShip,0,e.getBoardElement(rowNoShip,colNoShip));
+        Assert.assertEquals("Should have 17 ship elements.",17,e.getNumberOfShipElements());
+        Assert.assertEquals("Expected no ship at "+rowNoShip+","+colNoShip,0,e.getBoardElement(rowNoShip,colNoShip));
         e.fire(rowNoShip,colNoShip);
-        assertEquals("Expected a miss at "+rowNoShip+","+colNoShip,1,e.getBoardElement(rowNoShip,colNoShip));
-        assertEquals("Should have 17 ship elements.",17,e.getNumberOfShipElements());
+        Assert.assertEquals("Expected a miss at "+rowNoShip+","+colNoShip,1,e.getBoardElement(rowNoShip,colNoShip));
+        Assert.assertEquals("Should have 17 ship elements.",17,e.getNumberOfShipElements());
 
         /*
                  PLAYER BOARD
@@ -136,8 +140,8 @@ public class EnemyBoardTest {
         e.makeRandomBoard();
         e.fire(1,1);
 
-        assertEquals("Expected hasFired to be true for (1,1)", true, e.hasFired(1,1));
-        assertEquals("Expected hasFired to be false for (0,0)", false, e.hasFired(0,0));
+        Assert.assertEquals("Expected hasFired to be true for (1,1)", true, e.hasFired(1,1));
+        Assert.assertEquals("Expected hasFired to be false for (0,0)", false, e.hasFired(0,0));
 
     }
